Paul got err running stmts -- found hard-coded path for logo file in statement pgm (and invoices).  Added file cust_images to repos, and put basislogo.jpg in it.  Added two stbls to the .syn files.  +CUST_IMAGES points to the cust_images folder (using c:/apps... in the addon_dev.syn file, or ../apps in addon.syn).  +FORM_LOGO is set to basislogo.jpg. Cust can set this to whatever they want...
Added code to statements and invoices to derive the path/filename for the logo, and
use that in the setURL command instead of using the hard-coded path.

If the +CUST_IMAGES dir doesn't exist, or the +FORM_LOGO file doesn't exist, or either
stbl isn't set, program should still run, but w/ no logo.