Prop to branch. Original commit message:
Author: brucem
Date: 2010-03-04 10:50:06 -0700 (Thu, 04 Mar 2010)
New Revision: 6897

Modified:
  trunk/gl/src/glc_glpost.aon
Log:
process_id$, from the STBL  +PROCESS_ID, was coming in as null but needed to be padded to 16 for pointer read of glt_04 to work correctly.

This will eliminate missing Journal transactions when the same account number is used in multiple parts of the transaction being updated.

Bug # 4271

Needs to be put in branch also