Propagating mainline mods to branch.  Orig commit messages:
Author: jbrewer
Date: 2010-03-23 23:41:23 -0600 (Tue, 23 Mar 2010)
New Revision: 6947

Modified:
  trunk/op/src/opr_invoiceedit_o1.aon
Log:
Changed the mask for printing costs, and profits to use the Cost Mask. Also resized the columns to fit the Cost without it wrapping to fix bug 4367. Populate to the branch.
-----------------
Author: jbrewer
Date: 2010-03-26 00:00:46 -0600 (Fri, 26 Mar 2010)
New Revision: 6949

Modified:
  trunk/data/bar/ivm_itemwhse.xml
  trunk/data/cdf/ivm_itemwhse.cdf
Log:
Changed the way that the Location and Physical Inventory Code are disabled to allow the values to be displayed in every case to partially fix bug 4278. Populate to the branch.
------------------
Author: jbrewer
Date: 2010-03-26 14:13:10 -0600 (Fri, 26 Mar 2010)
New Revision: 6950

Modified:
  trunk/iv/src/ivu_physupdate.aon
Log:
Forced the Physical Selected flag to N every time a record is removed from ive-03 to fix bug 4278. Populate to the branch.
----------------
Author: jbrewer
Date: 2010-03-26 23:54:53 -0600 (Fri, 26 Mar 2010)
New Revision: 6951

Added:
  trunk/data/sync/ads_messages~iv_ser_one_zero.xml
Modified:
  trunk/data/bar/ive_count_entry.xml
  trunk/data/cdf/ive_count_entry.cdf
Log:
Force entry of serialized items to only allow quantity of 0 or 1 to work on bug 4280. Populate to the branch.