Propagate mods to branches.  Orig commit messages:
Author: jbrewer
Date: 2010-04-16 23:13:00 -0600 (Fri, 16 Apr 2010)
New Revision: 6998

Added:
  trunk/reports/VendorDetail_mast_fax.jasper
  trunk/reports/VendorDetail_mast_fax.jrxml
Modified:
  trunk/reports/VendorDetail.jasper
  trunk/reports/VendorDetail.jrxml
  trunk/reports/VendorDetail_cmts.jasper
Log:
Added Master Max and Email to the Vendor Detail report to work partially on bug 4370. Populate to the branch.
-----------
Author: jbrewer
Date: 2010-04-20 23:52:58 -0600 (Tue, 20 Apr 2010)
New Revision: 7000

Added:
  trunk/reports/VendorDetail_replenish.jasper
  trunk/reports/VendorDetail_replenish.jrxml
Modified:
  trunk/reports/VendorDetail.jasper
  trunk/reports/VendorDetail.jrxml
Log:
Added a new subform for the Replenishment Detail to the iReports Vendor Detail report to fix bug 4370. Populate to the branch.
----------
Author: jbrewer
Date: 2010-04-20 23:54:30 -0600 (Tue, 20 Apr 2010)
New Revision: 7001

Modified:
  trunk/data/bar/apm_vendrepl.xml
Log:
Corrected fields in the form to reflect actual fields. This is a result of fixing bug 4370 Populate to the branch!

Done 4/30.CH.