Prop recent fixes from mainline to 10.01 branch.  Orig commit messages: ----------------- Author: jbrewer Date: Thu Jul  1 15:45:58 2010 New Revision: 7133
Log:
Made changes to allow exiting the form if blank Lot/Serial Number is entered.

Modified:
   trunk/data/bar/ope_ordlsdet.xml
   trunk/data/cdf/ope_ordlsdet.cdf
---------------------
Author: jbrewer
Date: Thu Jul  1 23:09:24 2010
New Revision: 7134

Log:
Remove Commission Percent from Order and Invoice Entry forms.

Modified:
   trunk/data/bar/ope_invhdr.xml
   trunk/data/bar/ope_ordhdr.xml
   trunk/data/cdf/ope_invhdr.cdf
   trunk/data/cdf/ope_ordhdr.cdf
---------------------
Author: jbrewer
Date: Fri Jul  2 12:52:36 2010
New Revision: 7136

Log:
Set it so that if the Customer Number and Order Number are blank, all 3 buttons will be disabled. If the Customer Number is not blank, and the Order Number is, all 3 will light up. Once the Order Number is entered, the 2 Duplicate buttons are again disabled. This all to fix bug 4470.

Modified:
   trunk/data/bar/ope_invhdr.xml
   trunk/data/cdf/ope_invhdr.cdf