Added potential logic to limit the recordset of WOs to WOs flagged for Traveler printing (eg those listed in sfe_openedwo) when the SPROC is run from Traveler Printing (sfr_wotrav.aon).
The lead-ins pass new param REPORT_TYPE. The SPROC has new logic to wrap a sfe_openedwo-based JOIN around the pre-built select for travelers. BUT the .jasper has to be taught about the new param before the code can function. Code to handle the new param is currently REMd in SFHARDCOPY.prc.