Dashboard: Add a Manufacturing tab. Include an initial widget.
For ERs 7727 "AD - Dashboard: Add a manufacturing tab"
and     7728 "AD - Dashboard: Add a manufacturing tab widget to report
WO/SO linkages"

Note: WIP code for the tab and the initial widget, WOs/SOs grid widget,
was originally committed 5/2/2014 as part of commit for other Dashboard
mods. But the code there was disabled and incomplete. At that time, the
necessary adx_buildsproc.aon code was in place, and so is not committed
here.