Bug 9320 - AD - Rebuild FTI that runs as part of post-upgrade conversions can corrupt the table.
Moved the FTI rebuild to it's own separate auto-launch task that runs after the post-upgrade conversions are done and that task terminated.

There are Barista changes involved with this issue as well: altered bam_dd_export_FTI.bbj so that if any table that it's about to pass to bam_dd_export_table.bbj for drop/rebuild is open in the filesystem, it will be skipped and logged.