Reorder the callpoints in the CDF files in the old version so that they're in the same order as the callpoints in the CDF files of the new version. But do not cherry pick the commit where we reorder the CDFs to the new branch. Once the reordering is done, commit the changes and tag the commit.  Don't include the tagged commit when we cherry pick changes.  This whole reordering procedure will only work with non-Git-based upgrades, but hopefully it is not needed for Git-based upgrades.
Also, perform the reordering before we copy the user's files to the archive.  This means we need to copy the directory of files we reordered to a temporary directory called cdf_reordered.  Then we need to bring in the user's CDFs, then we need to use the CDFMassager's import mods utility to bring the user's CDFs into the reordered CDFs.  Once we do that, we can copy the cdf_reordered directory back to cdf and the cdf directory to cdf_old.