Bug 9975: In the customer payment portal, if EM rejects a user's new password, instead of crashing, it should display the error message to the user in the case that the password criteria has not been met.
Made changes to catch the error, and also adjusted some of the payment portal's behavior around changing the password on the first login.  Now, the portal will not let you bypass setting the password and going on to the customer payment form if you hit cancel.  You must set a new password before you go on to the payment form.  Part of the reason for this is because whether or not you've set your password is determined by your last use date.  I've moved setting that date to when the user arrives at the payment form.