Bug 10132: We should verify whether GIT_AUTH_IDs are forks/clones of the official repo at the time they're added to the ADX_GIT_AUTH table.
Introduce a BWRI callpoint to ADX_GIT_AUTH that validates the Repo being added before saving the record.
Add an additional constructor and obtainURLAndCredentials method signature that accept the fields of a GIT_AUTH_ID record as parameters, so the record does not have to be in the database already to check the validity of the site. Don't write the record to the ADX_GIT_AUTH table if the repo isn't valid.